Add missing tab to the navigation and improve UX
Add missing tab to the navigation and improve UX

Problem:
=======
Our current navigation bar was missing the crucial “help” dropdown menu that includes both support and contact links.  And the mouseover behavior was causing the menu to flash on certain browsers.

[Add missing tab story in Pivotal Tracker](http://google.com)

Solution:
=========
We solved this by first adding the ‘help’ dropdown with contact and support links to the current nav bar.  Then we refactored some of the CSS and changed the interaction to be a click (to get the dropdown to show) instead of a hover.  We discussed this change with the design team.

Changes:
========
Added new `navigation.css` file to compartmentalize the styles for the nav
Updated `navigation.html` partial to include the new help tab
Added view tests to confirm

Screenshots:
============